x11-themes/adwaita-icon-theme: Fix logic to remove old directory
Thanks-to: Agostino Sarubbo
Thanks-to: Joonas Niilola
Closes: https://bugs.gentoo.org/847877

+++ x11-themes/adwaita-icon-theme/adwaita-icon-theme-42.0_p2.ebuild
@@ -67,7 +67,9 @@ src_compile() {
 
 pkg_preinst() {
 	# Needed until bug #834600 is solved
-	rm -r "${EROOT}"/usr/share/cursors/xorg-x11/Adwaita || die
+	if [[ -d "${EROOT}"/usr/share/cursors/xorg-x11/Adwaita ]] ; then
+		rm -r "${EROOT}"/usr/share/cursors/xorg-x11/Adwaita || die
+	fi
 }
 
 src_install() {
